rename gpu processor package to gpuattributes and address comments
