+++ plugins/processors/gpuattributes/config_test.go
@@ -1,7 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: MIT
 
-package gpu
+package gpuattributes
 
 import (
 	"testing"
