+++ translator/translate/otel/processor/gpu/translator.go
@@ -8,7 +8,7 @@ import (
 	"go.opentelemetry.io/collector/confmap"
 	"go.opentelemetry.io/collector/processor"
 
-	"github.com/aws/amazon-cloudwatch-agent/plugins/processors/gpu"
+	"github.com/aws/amazon-cloudwatch-agent/plugins/processors/gpuattributes"
 	"github.com/aws/amazon-cloudwatch-agent/translator/translate/otel/common"
 )
 
@@ -20,7 +20,7 @@ type translator struct {
 var _ common.Translator[component.Config] = (*translator)(nil)
 
 func NewTranslatorWithName(name string) common.Translator[component.Config] {
-	return &translator{name, gpu.NewFactory()}
+	return &translator{name, gpuattributes.NewFactory()}
 }
 
 func (t *translator) ID() component.ID {
@@ -28,6 +28,6 @@ func (t *translator) ID() component.ID {
 }
 
 func (t *translator) Translate(conf *confmap.Conf) (component.Config, error) {
-	cfg := t.factory.CreateDefaultConfig().(*gpu.Config)
+	cfg := t.factory.CreateDefaultConfig().(*gpuattributes.Config)
 	return cfg, nil
 }
